Move access token text in user list to under the dropdown
